Factor out a `test_scheduler` module
